+++ test/dockerValidator.test.ts
@@ -506,6 +506,17 @@ function assertOnbuildTriggerDisallowed(diagnostic: Diagnostic, trigger: string,
     assert.equal(diagnostic.range.end.character, endCharacter);
 }
 
+function assertDirectiveEscapeDuplicated(diagnostic: Diagnostic, startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
+    assert.equal(diagnostic.code, ValidationCode.DUPLICATED_ESCAPE_DIRECTIVE);
+    assert.equal(diagnostic.severity, DiagnosticSeverity.Error);
+    assert.equal(diagnostic.source, source);
+    assert.equal(diagnostic.message, Validator.getDiagnosticMessage_DirectiveEscapeDuplicated());
+    assert.equal(diagnostic.range.start.line, startLine);
+    assert.equal(diagnostic.range.start.character, startCharacter);
+    assert.equal(diagnostic.range.end.line, endLine);
+    assert.equal(diagnostic.range.end.character, endCharacter);
+}
+
 function assertDirectiveEscapeInvalid(diagnostic: Diagnostic, value: string, startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
     assert.equal(diagnostic.code, ValidationCode.INVALID_ESCAPE_DIRECTIVE);
     assert.equal(diagnostic.severity, DiagnosticSeverity.Error);
@@ -1703,6 +1714,13 @@ describe("Docker Validator Tests", function() {
             it("ignored on second line", function() {
                 let diagnostics = validateDockerfile("\n# escape=a\nFROM node");
                 assert.equal(diagnostics.length, 0);
+            });
+
+            it("duplicated escape directive", function() {
+                let diagnostics = validateDockerfile("# escape=`\n# escape=`\nFROM alpine");
+                assert.equal(diagnostics.length, 2);
+                assertDirectiveEscapeDuplicated(diagnostics[0], 0, 2, 0, 10);
+                assertDirectiveEscapeDuplicated(diagnostics[1], 1, 2, 1, 10);
             });
         });